requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0
python-dotenv>=1.0.0
gitpython>=3.1.40
pydantic>=2.5.0
langsmith>=0.0.72
